requirements.txt
azure-functions
yfinance
pandas
sqlalchemy
pyodbc
tenacity
azure-keyvault-secrets
azure-identity
python-dotenv